requirements.txt
certifi==2016.9.26
future==0.15.2
pony==0.7
python-telegram-bot==5.1.1
urllib3==1.18
